Nav-bar improvements and projects section test
Here I've made the nav-bar sticky so the nav bar is still accessible after scrolling down the page. I have also given the links a highlight style effect using a transition. Along with this, after clicking the links, the page now scrolls to the desired section as apposed to 'snapping' to it.

With the projects section I have added the same project twice, just to see the layout of the page with 3 projects on desktop. I'll be replacing these with better projects as and when I build them. For now I'm just testing the layout.